feat: Increase SSDP discovery timeout
Increased the timeout for the SSDP discovery socket from 10 to 100 seconds.

This change allows for more time for UPnP devices on the network to respond, which can be useful in environments with slower or more congested networks. A longer timeout reduces the chance of missing devices that are slow to respond to the discovery message.